+++ src/main.rs
@@ -2,6 +2,7 @@ mod admin_client;
 mod build_info;
 mod config;
 mod routes;
+mod session_client;
 mod telemetry;
 
 use std::sync::Arc;
@@ -12,6 +13,7 @@ use axum::Router;
 use axum_prometheus::PrometheusMetricLayer;
 use build_info::BuildInfo;
 use config::Config;
+use session_client::SessionClient;
 use tower_http::services::ServeDir;
 use tower_http::trace::TraceLayer;
 
@@ -19,6 +21,7 @@ pub struct AppState {
     config: Config,
     build_info: BuildInfo,
     admin_client: AdminClient,
+    session_client: SessionClient,
 }
 
 #[tokio::main]
@@ -40,10 +43,12 @@ async fn main() {
     );
 
     let admin_client = AdminClient::new(config.admin_api_url.clone());
+    let session_client = SessionClient::new(config.redis_host.clone(), config.redis_port).await;
     let state = Arc::new(AppState {
         config,
         build_info,
         admin_client,
+        session_client,
     });
 
     let (prometheus_layer, metric_handle) = PrometheusMetricLayer::pair();